chat_with_iit-main/requirements.txt
openai
numpy
tiktoken
psycopg2-binary
streamlit
requests